fix(core): Restore original property descriptor on unregister
Fixes a potential  "Maximum callstack size exceeded" error. The current implementation doesn't cleanup the defineProperty call on the element. For every register() call, a new object descriptor is created which calls the previous one and so on. After a lot of calls of register() and unregister(), the JS engine will finally throw a callstack size error.